Move vector-owned scheduling symbols
